week 3/customer-support-response-generator/requirements.txt
opopenai>=2.16.0
python-dotenv>=1.0.0
